fix: resolve remaining build issues
- Fix duplicate model definitions in places domain
- Run go mod tidy to ensure all dependencies are properly resolved
- Remove backup files
